# Don't compile regex in is_valid_account_id