--- a/core/primitives/src/utils.rs
+++ b/core/primitives/src/utils.rs
@@ -2,6 +2,7 @@ use bs58;
 use byteorder::{LittleEndian, WriteBytesExt};
 use protobuf::{well_known_types::StringValue, SingularPtrField};
 use std::convert::{TryFrom, TryInto};
+use lazy_static::lazy_static;
 
 use crate::types::{AccountId, ShardId};
 use regex::Regex;
@@ -22,9 +23,12 @@ pub fn bs58_vec2str(buf: &[u8]) -> String {
     bs58::encode(buf).into_string()
 }
 
+lazy_static! {
+    static ref VALID_ACCOUNT_ID: Regex = Regex::new(r"^[a-z0-9@._\-]{5,32}$").unwrap();
+}
+
 pub fn is_valid_account_id(account_id: &AccountId) -> bool {
-    let re = Regex::new(r"^[a-z0-9@._\-]{5,32}$").unwrap();
-    re.is_match(account_id)
+    VALID_ACCOUNT_ID.is_match(account_id)
 }
 
 pub fn to_string_value(s: String) -> StringValue {