before:
```
test runtime_send_money                     ... bench:     292,390 ns/iter (+/- 44,812)
test runtime_wasm_bad_code                  ... bench:   1,183,672 ns/iter (+/- 118,586)
test runtime_wasm_benchmark_10_reads_legacy ... bench:   1,013,094 ns/iter (+/- 133,674)
test runtime_wasm_benchmark_storage_100     ... bench:   1,922,518 ns/iter (+/- 223,659)
test runtime_wasm_benchmark_storage_1000    ... bench:  12,751,034 ns/iter (+/- 1,571,115)
test runtime_wasm_benchmark_sum_1000        ... bench:     913,720 ns/iter (+/- 278,751)
test runtime_wasm_benchmark_sum_1000000     ... bench:  13,472,694 ns/iter (+/- 988,551)
test runtime_wasm_set_value                 ... bench:     901,383 ns/iter (+/- 99,364)
```
after:
```
test runtime_send_money                     ... bench:     162,952 ns/iter (+/- 27,979)
test runtime_wasm_bad_code                  ... bench:   1,037,134 ns/iter (+/- 221,946)
test runtime_wasm_benchmark_10_reads_legacy ... bench:     883,843 ns/iter (+/- 90,748)
test runtime_wasm_benchmark_storage_100     ... bench:   1,779,139 ns/iter (+/- 371,843)
test runtime_wasm_benchmark_storage_1000    ... bench:  12,172,454 ns/iter (+/- 1,912,120)
test runtime_wasm_benchmark_sum_1000        ... bench:     761,535 ns/iter (+/- 123,590)
test runtime_wasm_benchmark_sum_1000000     ... bench:  13,315,627 ns/iter (+/- 1,206,112)
test runtime_wasm_set_value                 ... bench:     767,202 ns/iter (+/- 146,407)
```